Add tests for UNIX sockets
Extend functions in `test/helpers/macros.js` to accept either a socket
path or a port.

In assertListen and assertConnect verify that the port or socket that's
emitted when hook::listening or hook::connected fires matches the port or
socket we were attempting to connect to. (Otherwise, these tests can appear to
pass even when hook-socket support is unimplemented because the hook-socket was
just ignored and the hooks would connect on the default port instead).

This functionality requires the libuv patches mentioned here:

https://github.com/joyent/libuv/issues/158